% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColonyTrack.R
\docType{package}
\name{ColonyTrack}
\alias{ColonyTrack}
\title{Analysis of Tracking Data From the PhenoSys ColonyRack Cage.}
\description{
This package prepares raw timestamped Radio Frequency ID data into a standardised storage format, infers the position of each subject at each timepoint and extracts behavioural metrics.
}
\section{Functions provided}{


  \code{\link{read_data}}

  \code{\link{calculate_metrics}}

  \code{\link{combine_metrics}}

  \code{\link{get_metrics}}

  \code{\link{plot_metric}}

  \code{\link{plot_ethogram}}
}

