% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_metrics.R
\name{combine_metrics}
\alias{combine_metrics}
\title{Combine metrics objects.}
\usage{
combine_metrics(metrics.list)
}
\arguments{
\item{metrics.list}{A list containing any number of
\code{colonytrack_metrics} objects in any order. These will be merged into
one combined \code{colonytrack_metrics} object sorted by date. This
parameter may also be a character list of filenames pointing to saved RData
archives containing the \code{colonytrack_metrics} objects.}
}
\value{
A \code{colonytrack_metrics} object.
}
\description{
Joins metrics results into one contiguous metrics object.
}
\details{
This function takes processed results from several
\code{\link{calculate_metrics}} runs and merges them into one
\code{colonytrack_metrics} object. This allows metrics to be calculated in
batches to limit memory usage.
}
\seealso{
\code{\link{calculate_metrics}}.
}
