% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics.R
\name{get_metrics}
\alias{get_metrics}
\title{Retrieve one or more metrics.}
\usage{
get_metrics(subject, metrics, days = "all")
}
\arguments{
\item{subject}{The name of the subject for which the data is to be retrieved.}

\item{metrics}{A \code{colonytrack_metrics} object, as returned by
\code{\link{calculate_metrics}}.}

\item{days}{Specify the days for which metrics should be retrieved. This
may be in any of the standard formats (names, indices, or a logical vector).
Default is to include all days present in the metrics object.}
}
\value{
Returns a matrix with days/nights as rows and metrics as columns.
}
\description{
Retrieves data for all individual metrics for a single subject.
}
