% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metric.R
\name{plot_metric}
\alias{plot_metric}
\title{Plot one or more metrics.}
\usage{
plot_metric(
  metric,
  metrics,
  days = "all",
  subjects = "all",
  scheme = "rainbow",
  col = NULL,
  file = NULL,
  cex.axis = 1,
  las = 2,
  mar = c(8, 8, 4, 0),
  plot = TRUE
)
}
\arguments{
\item{metric}{The name of the metric that is to be plotted.}

\item{metrics}{A \code{colonytrack_metrics} object, as returned by
\code{\link{calculate_metrics}}.}

\item{days}{Specify the days to be included is the metrics calculation. This
may be in any of the standard formats (names, indices, or a logical vector).
Default is to include all days present in the input data file.}

\item{subjects}{Specify the subjects to be included is the metrics
calculation. This may be in any of the standard formats (names, indices, or
a logical vector). Default is to include all subjects present in the input
data file. If a vector of subjects is supplied, the resulting plot will
respect the ordering of this vector.}

\item{scheme}{A code specifying the colour scheme to be used. See Details for
information on the available schemes. Default is 'rainbow'.}

\item{col}{A named vector of colours. This vector must contain the SubjectIDs
as names. Any subjects not present in this vector will be drawn in black. If
not \code{NULL}, then this parameter overrides \code{scheme}. Default is
\code{NULL}.}

\item{file}{An optional filename for PDF output. If supplied, the plot will be
written to a PDF document with this name. The file is saved relative to the
working directory. The default is to plot to the active graphics device.}

\item{cex.axis}{The scaling factor for the axis labels (see
\code{\link{par}}). Default (0.75) can be overridden for different sized
plot devices.}

\item{las}{The label aspect (see \code{\link{par}}. Can be overridden based on
the user's preferences.}

\item{mar}{The margins surrounding the plot (see \code{\link{par}}). Can be
overridden for different length ID labels or to match different sized plot
devices.}

\item{plot}{If \code{FALSE}, then the plot is not drawn and only the plotting
data are returned.}
}
\description{
Plots a line graph of the individual metrics over time.
}
\details{
The data used in plotting (the plotted values for each subject and the
colours) are returned invisibly to enable further plot customisation.
}
