% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Preparation of tracking data.}
\usage{
read_data(
  dataFiles,
  subjectFile,
  networkFile,
  eventsFile,
  cageQualityFile = NULL
)
}
\arguments{
\item{dataFiles}{A vector of filenames. Each file is a CSV dump of antenna
contact data. These files should cover a contiguous period encompassing the
whole experiment. Files can be given as full pathnames so that the often
large data volumes can be stored outside the analysis directory (for
example on an external fileserver). Any file format readable by
\code{\link[utils]{read.table}} can be used.}

\item{subjectFile}{A tab-delimited file describing the subjects/animals. This
information is required to link RFID tags to variables/factors such as
experimental group, age, weight etc.}

\item{networkFile}{A tab-delimited file describing the layout of the cage
system. This is currently in the the form of an unweighted undirected
network.}

\item{eventsFile}{A tab-delimited file describing events. This must minimally
include details of the light cycle. See Details for more information.}

\item{cageQualityFile}{An optional tab-delimited file describing the content
of the cages (food, bedding, toys etc.). See Details for more information.}
}
\value{
A \code{colonytrack_data} object.
}
\description{
Reads raw antenna contact data and metadata files.
}
\details{
This function requires raw CSV data from the Colony Cage as well as a
subject/animal description file, a cage layout description file and a log of
events. An optional metadata file allows finer description of the cages.

The processed data are padded to start at "lights on" of the first day to
"lights off" of the last day (as specified in the 'Events' file).
Nevertheless, the first and last days (and any interruptions in recording)
will result in missing data that may need to be handled during analysis. The
actogram (produced by \code{\link{plot}} of the \code{colonytrack_data}
object) can be used to identify such missing blocks.
}
